Backend/flask/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mysql-connector-python==8.0.32
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pika==1.3.1
Pillow==9.5.0
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.4
protobuf==3.20.3
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.4
requests==2.28.2
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post4
threadpoolctl==3.1.0
tldextract==3.4.0
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
zope.event==4.6
zope.interface==6.0
